ACCUMULO-259 - Merging up to date
git-svn-id: https://svn.apache.org/repos/asf/accumulo/branches/ACCUMULO-259@1433166 13f79535-47bb-0310-9956-ffa450edef68

+++ server/src/main/java/org/apache/accumulo/server/test/NativeMapConcurrencyTest.java
@@ -25,6 +25,9 @@ import org.apache.accumulo.core.data.Value;
 import org.apache.accumulo.server.tabletserver.NativeMap;
 import org.apache.hadoop.io.Text;
 
+import com.beust.jcommander.JCommander;
+import com.beust.jcommander.Parameter;
+
 public class NativeMapConcurrencyTest {
   
   private static final byte ROW_PREFIX[] = new byte[] {'r'};
@@ -70,14 +73,30 @@ public class NativeMapConcurrencyTest {
     return num / (ms / 1000.0);
   }
   
+  static class Opts {
+    @Parameter(names="--rows", description="rows", required = true)
+    int rows = 0;
+    @Parameter(names="--cols", description="cols")
+    int cols = 1;
+    @Parameter(names="--threads", description="threads")
+    int threads = 1;
+    @Parameter(names="--writeThreads", description="write threads")
+    int writeThreads = 1;
+    @Parameter(names="-help", help=true)
+    boolean help = false;
+  }
+  
   public static void main(String[] args) {
-    int rows = Integer.parseInt(args[0]);
-    int cols = Integer.parseInt(args[1]);
-    int threads = Integer.parseInt(args[2]);
-    int writeThreads = Integer.parseInt(args[3]);
-    
-    NativeMap nm = create(rows, cols);
-    runTest(nm, rows, cols, threads, writeThreads);
+    Opts opts = new Opts();
+    JCommander jc = new JCommander(opts);
+    jc.setProgramName(NativeMapConcurrencyTest.class.getName());
+    jc.parse(args);
+    if (opts.help) {
+      jc.usage();
+      return;
+    }
+    NativeMap nm = create(opts.rows, opts.cols);
+    runTest(nm, opts.rows, opts.cols, opts.threads, opts.writeThreads);
     nm.delete();
   }
